requirements.txt
Django==1.10.5
djangorestframework==3.6.2
easy_install mysql-python